+++ api/Controllers/TrainingProgramController.cs
@@ -3,7 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using api.Dtos.TrainingProgram;
+using api.Extensions;
 using api.Interfaces;
+using api.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,29 +18,39 @@ namespace api.Controllers
     public class TrainingProgramController : ControllerBase
     {
         private readonly ITrainingProgramRepository _programRepository;
-        private readonly IUserRepository _userRepository;
+        private readonly UserManager<User> _userManager;
         private readonly IExerciseRepository _exerciseRepository;
-        public TrainingProgramController(ITrainingProgramRepository program_repository, IUserRepository user_repository, IExerciseRepository exercise_repository)
+
+        public TrainingProgramController(IExerciseRepository exercise_repository, ITrainingProgramRepository program_repository, UserManager<User> userManager)
         {
-            _programRepository = program_repository;
-            _userRepository = user_repository;
             _exerciseRepository = exercise_repository;
+            _programRepository = program_repository;
+            _userManager = userManager;
         }
-        // TODO authenticate
-        // TODO redo this to use authentication to determine user id
         [HttpGet]
-        public async Task<IActionResult> GetTrainingProgramsForUser(string userId)
+        [Authorize]
+        public async Task<IActionResult> GetTrainingProgramsForUser()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
             var programs = await _programRepository.GetTrainingProgramsForUser(userId);
+            if (programs == null || !programs.Any())
+            {
+                return NotFound();
+            }
             return Ok(programs);
         }
 
-        // TODO authenticate
         [HttpGet("{programId}")]
+        [Authorize]
         public async Task<IActionResult> GetTrainingProgramById([FromRoute] int programId)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
             var program = await _programRepository.GetTrainingProgramById(programId);
-            if (program == null)
+            if (program == null || program.UserId != userId)
             {
                 return NotFound();
             }
@@ -44,10 +58,14 @@ namespace api.Controllers
         }
 
         [HttpGet("program/{programId}/days")]
+        [Authorize]
         public async Task<IActionResult> GetDaysByProgramId([FromRoute] int programId)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
             var days = await _programRepository.GetDaysByProgramId(programId);
-            if (days == null || !days.Any())
+            if (days == null || !days.Any() || days.First().TrainingProgram.UserId != userId)
             {
                 return NotFound();
             }
@@ -55,10 +73,14 @@ namespace api.Controllers
         }
 
         [HttpGet("day/{dayId}")]
+        [Authorize]
         public async Task<IActionResult> GetDayById([FromRoute] int dayId)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
             var day = await _programRepository.GetDayById(dayId);
-            if (day == null)
+            if (day == null || day.TrainingProgram.UserId != userId)
             {
                 return NotFound();
             }
@@ -68,6 +90,13 @@ namespace api.Controllers
         [HttpGet("day/{dayId}/exercises")]
         public async Task<IActionResult> GetExercisesByDay([FromRoute] int dayId)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.ProgramDayExists(userId, dayId))
+            {
+                return NotFound();
+            }
             var exercises = await _programRepository.GetExercisesByDay(dayId);
             if (exercises == null || !exercises.Any())
             {
@@ -77,8 +106,16 @@ namespace api.Controllers
         }
 
         [HttpGet("exercise/{id}")]
+        [Authorize]
         public async Task<IActionResult> GetExerciseById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.ProgrammedExerciseExists(userId, id))
+            {
+                return NotFound();
+            }
             var exercise = await _programRepository.GetExerciseById(id);
             if (exercise == null)
             {
@@ -88,23 +125,28 @@ namespace api.Controllers
         }
 
         [HttpPost("create")]
+        [Authorize]
         public async Task<IActionResult> CreateTrainingProgram([FromBody] TrainingProgramCreateDto programDto)
         {
-            if (!await _userRepository.UserExists(programDto.UserId))
-            {
-                return BadRequest("User does not exist.");
-            }
-            var newProgram = await _programRepository.CreateTrainingProgram(programDto);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            var newProgram = await _programRepository.CreateTrainingProgram(userId, programDto);
             if (newProgram == null)
             {
                 return BadRequest("Could not create training program.");
             }
             return CreatedAtAction(nameof(GetTrainingProgramById), new { userId = newProgram.UserId, programId = newProgram.Id }, newProgram);
         }
+
         [HttpPost("day/create")]
+        [Authorize]
         public async Task<IActionResult> CreateProgramDay([FromBody] ProgramDayCreateDto programDayDto)
         {
-            if (!await _programRepository.TrainingProgramExists(programDayDto.TrainingProgramId))
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.TrainingProgramExists(userId, programDayDto.TrainingProgramId))
             {
                 return BadRequest("Training program does not exist.");
             }
@@ -117,15 +159,18 @@ namespace api.Controllers
         }
 
         [HttpPost("exercise/create")]
+        [Authorize]
         public async Task<IActionResult> CreateProgrammedExercise([FromBody] ProgrammedExerciseCreateDto exerciseDto)
         {
-            if (!await _userRepository.UserExists(exerciseDto.UserId))
-                return BadRequest("User does not exist.");
-            if (!await _programRepository.ProgramDayExists(exerciseDto.ProgramDayId))
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+
+            if (!await _programRepository.ProgramDayExists(userId, exerciseDto.ProgramDayId))
                 return BadRequest("Program day does not exist.");
-            if (!await _exerciseRepository.ExerciseExists(exerciseDto.ExerciseId))
+            if (!await _exerciseRepository.ExerciseExists(userId, exerciseDto.ExerciseId))
                 return BadRequest("Exercise does not exist.");
-            if (await _programRepository.ProgrammedExercisePositionExists(exerciseDto.ProgramDayId, exerciseDto.Position))
+            if (await _programRepository.ProgrammedExercisePositionExists(userId, exerciseDto.ProgramDayId, exerciseDto.Position))
                 return BadRequest("An exercise already exists at this position in the program day.");
             var newExercise = await _programRepository.CreateProgrammedExercise(exerciseDto);
             if (newExercise == null)
@@ -135,13 +180,19 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetExerciseById), new { id = newExercise.Id }, newExercise);
         }
 
-        // TODO authenticate
-        // TODO redo this to use authentication to determine user id
         [HttpPut("update/{programId}")]
+        [Authorize]
         public async Task<IActionResult> UpdateTrainingProgram([FromRoute] int programId, [FromBody] TrainingProgramUpdateDto programDto)
         {
-            var existing_program = await _programRepository.UpdateTrainingProgram(programId, programDto);
-            if (existing_program == null)
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.TrainingProgramExists(userId, programId))
+            {
+                return NotFound();
+            }
+            var updated_program = await _programRepository.UpdateTrainingProgram(programId, programDto);
+            if (updated_program == null)
             {
                 return NotFound();
             }
@@ -149,8 +200,16 @@ namespace api.Controllers
         }
 
         [HttpPut("update/day/{dayId}")]
+        [Authorize]
         public async Task<IActionResult> UpdateProgramDay([FromRoute] int dayId, [FromBody] ProgramDayUpdateDto dayDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.ProgramDayExists(userId, dayId))
+            {
+                return NotFound();
+            }
             var existingDay = await _programRepository.UpdateProgramDay(dayId, dayDto);
             if (existingDay == null)
             {
@@ -160,8 +219,16 @@ namespace api.Controllers
         }
 
         [HttpPut("update/exercise/{id}")]
+        [Authorize]
         public async Task<IActionResult> UpdateProgrammedExercise([FromRoute] int id, [FromBody] ProgrammedExerciseUpdateDto exerciseDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.ProgrammedExerciseExists(userId, id))
+            {
+                return NotFound();
+            }
             var existingExercise = await _programRepository.UpdateProgrammedExercise(id, exerciseDto);
             if (existingExercise == null)
             {
@@ -170,13 +237,14 @@ namespace api.Controllers
             return NoContent();
         }
 
-
-        // TODO authenticate
-        // TODO redo this to use authentication to determine user id
         [HttpDelete("delete")]
+        [Authorize]
         public async Task<IActionResult> DeleteTrainingProgram(int programId)
         {
-            if (!await _programRepository.TrainingProgramExists(programId))
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.TrainingProgramExists(userId, programId))
             {
                 return NotFound();
             }
@@ -193,32 +261,32 @@ namespace api.Controllers
             return NoContent();
         }
 
-        // TODO authenticate
-        // TODO redo this to use authentication to determine user id
         [HttpDelete("delete/day")]
+        [Authorize]
         public async Task<IActionResult> DeleteProgramDay(int dayId)
         {
-            if (!await _programRepository.ProgramDayExists(dayId))
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.ProgramDayExists(userId, dayId))
             {
                 return NotFound();
-            }
-            var exercises = await _programRepository.GetExercisesByDay(dayId);
-            if (exercises != null)
-            {
-                foreach (var exercise in exercises)
-                {
-                    if (exercise == null) continue;
-                    await _programRepository.DeleteProgrammedExercise(exercise.Id);
-                }
             }
             await _programRepository.DeleteProgramDay(dayId);
             return NoContent();
         }
-        
-        // TODO authenticate
+
         [HttpDelete("delete/exercise/")]
+        [Authorize]
         public async Task<IActionResult> DeleteProgrammedExercise(int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var userId = User.GetId();
+            if (!await _programRepository.ProgrammedExerciseExists(userId, id))
+            {
+                return NotFound();
+            }
             var existingExercise = await _programRepository.DeleteProgrammedExercise(id);
             if (existingExercise == null)
             {
